Removing part of the page due to unseen error
Google is telling me that there are tap targets that are too close
together, despite me only having 1 interactive object on the page. It
says that "<div></div>" is too close to something, yet the only div on
the page is the one containing the logo. I have removed the div, and
will now see how it works.